Remove use of the `f` filter plugin in roles
Guess I gave up on that filter at some point? Probably when I found out
it would need to be `nansi.util.f`, 'cause that kinda ruins it.

For posterity's sake, the `f` filter was an attempt at a "general" variable
formatter. I don't think it ever got that interesting or useful in it's
lifetime.